pyproject.toml
[tool.poetry]
name = "pyweb-hw11"
version = "0.1.0"
description = ""
authors = ["shuguruiua <popov.yk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.0"
sqlalchemy = "2.0.28"
alembic = "1.13.1"
asyncpg = "0.29.0"
uvicorn = "0.29.0"
pydantic = {extras = ["email"], version = "2.6.4"}
python-multipart = "0.0.9"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
libgravatar = "1.0.4"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
bcrypt = "4.0.01"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
